pyproject.toml
[tool.black]
exclude = '''
/(
    \.mypy_cache
  | \.pytest_cache
  | migrations
)/
'''
line-length = 119

[tool.isort]
profile = "black"
skip_glob = [".mypy_cache", ".pytest_cache", "migrations"]

[tool.mypy]
plugins = ["pydantic.mypy", "mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "core.settings"

[tool.coverage]
[tool.coverage.run]
omit = ["*migrations*", "*asgi.py", "*wsgi.py", "*manage.py", "*conftest.py"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
addopts = "--reuse-db"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[tool.poetry]
authors = ["Evgeni Sautin <evgeni.sautin@gmail.com>"]
description = "Find the skills you need to get a job in IT"
license = "MIT"
name = "SkillHunter"
version = "3.0.0"

[tool.poetry.dependencies]
Django = "^4.0.1"
Faker = "^11.3.0"
Jinja2 = "^3.0.3"
aiohttp = "^3.8.1"
beautifulsoup4 = "^4.10.0"
boto3 = "^1.20.42"
celery = {extras = ["pytest"], version = "^5.2.3"}
django-csp = "^3.7"
django-debug-toolbar = "^3.2.4"
django-linear-migrations = "^2.3.0"
django-otp = "^1.1.3"
django-permissions-policy = "^4.7.0"
django-storages = "^1.12.3"
djangorestframework = "^3.13.1"
drf-spectacular = "^0.21.1"
flashtext = "^2.7"
gunicorn = "^20.1.0"
pdftotext = "^2.2.2"
psycopg2-binary = "^2.9.3"
python = "^3.10"
python-multipart = "^0.0.5"
qrcode = "^7.3.1"
redis = {extras = ["hiredis"], version = "^4.1.2"}
requests = "^2.27.1"
selenium = "^4.1.0"
sentry-sdk = "^1.5.3"
uvicorn = {extras = ["standard"], version = "^0.17.1"}

[tool.poetry.dev-dependencies]
black = "^21.12b0"
django-stubs = "^1.9.0"
flake8 = "^4.0.1"
ipython = "^8.0.0"
isort = "^5.10.1"
mypy = "^0.930"
pre-commit = "^2.16.0"
pytest-asyncio = "^0.16.0"
pytest-cov = {extras = ["toml"], version = "^3.0.0"}
pytest-django = "^4.5.2"
pytest-xdist = {extras = ["psutil"], version = "^2.5.0"}

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
